requirements.txt
torch
torchvision
opencv-python
google-genai
fastapi
uvicorn
numpy
pandas
furhat-remote-api